Drawing samples from the prior distribution given by generalised linear model.

This function provides the inputs for mcmc sampler. Samples are drawn from the prior distribution of the sampling rate and the corresponding densities are evaluated. The prior distribution is informative and given by a posterior distribution inferred from Bayesian hierarchical model using empirical sampling rate and group covariates.

samples.from.GLM.prior(df.sampling,dc,iteration,sample.method,glm.model,seed)

Arguments

df.sampling
a list of data tables each of which includes data used for infering the distribution of sampling rates through a hierarchical model, i.e. the number of eligible individuals and individual sampled in each subpopulation, the covariates of individuals in a subpopulation.

dc
a data table includes transmission counts between subpopulations and covariates of transmitters' and recipients' subpopulations.

iteration
the number of iterations for mcmc sampler, which decides the number of prior samples needed.

sample.method
the methods for evaluating density of each samples. method='empirical' indicates density of samples come from an empirical distribution, method='betaapprox' indicates a beta distribution is used to approximate the distribution of samples.

glm.model
the type of glm model for infering the sampling rates. glm.model='binomial' indicates using binomial model, glm.model='beta-binomial' indicates using beta-binomial model. Choices of glm.model needed to be determined by posterior predictive checks.

seed
the random seed.

Examples

library(data.table)
library(Boom)

#> Loading required package: MASS

#> 
#> Attaching package: ‘Boom’

#> The following object is masked from ‘package:coda’:
#> 
#>     thin

#> The following object is masked from ‘package:stats’:
#> 
#>     rWishart

set.seed(123)
# assume 120 infected individuals in 1 and 80 infected individuals in 2
# 70% are female in 1 and 60% are females in 2
# in 1 F: 25% 15-24 60% 25-34 15% 35+
# in 1 M: 20% 15-24 50% 25-34 30% 35+
# in 2 F: 25% 15-24 50% 25-34 25% 35+
# in 2 M: 20% 15-24 40% 25-34 40% 35+
dg<-data.table(COMM_NUM_B=c(rep(1,120),rep(2,80)),INF=rep(1,200))
num.f.1<-round(120*0.7)
num.m.1<-120-round(120*0.7)
num.f.2<-round(80*0.6)
num.m.2<-80-round(80*0.6)

dg$SEX<-c(rep('F',num.f.1),
          rep('M',num.m.1),
          rep('F',num.f.2),
          rep('M',num.m.2))
dg$AGE_AT_MID_C<-c(rep('15-24',round(num.f.1*0.25)),
                   rep('25-34',round(num.f.1*0.6)),
                   rep('35+',num.f.1-round(num.f.1*0.25)-round(num.f.1*0.6)),
                   rep('15-24',round(num.m.1*0.2)),
                   rep('25-34',round(num.m.1*0.5)),
                   rep('35+',num.m.1-round(num.m.1*0.2)-round(num.m.1*0.5)),
                   rep('15-24',round(num.f.2*0.25)),
                   rep('25-34',round(num.f.2*0.5)),
                   rep('35+',num.f.2-round(num.f.2*0.25)-round(num.f.2*0.5)),
                   rep('15-24',round(num.m.2*0.2)),
                   rep('25-34',round(num.m.2*0.4)),
                   rep('35+',num.m.2-round(num.m.2*0.2)-round(num.m.2*0.4))
                   )

# regression coefficients
dg              <- dg[, list(TRIAL=length(INF)),by=c('AGE_AT_MID_C','SEX','COMM_NUM_B')]
comm<-c(0.5,-0.5)
a<-0.4
male<-0.4
young<-0.3
midage<-0.2

#	binarize age, sex
dg[, AGE1:= as.integer(AGE_AT_MID_C=='15-24')]

#>     AGE_AT_MID_C SEX COMM_NUM_B TRIAL AGE1
#>  1:        15-24   F          1    21    1
#>  2:        25-34   F          1    50    0
#>  3:          35+   F          1    13    0
#>  4:        15-24   M          1     7    1
#>  5:        25-34   M          1    18    0
#>  6:          35+   M          1    11    0
#>  7:        15-24   F          2    12    1
#>  8:        25-34   F          2    24    0
#>  9:          35+   F          2    12    0
#> 10:        15-24   M          2     6    1
#> 11:        25-34   M          2    13    0
#> 12:          35+   M          2    13    0

dg[, AGE2:= as.integer(AGE_AT_MID_C=='25-34')]

#>     AGE_AT_MID_C SEX COMM_NUM_B TRIAL AGE1 AGE2
#>  1:        15-24   F          1    21    1    0
#>  2:        25-34   F          1    50    0    1
#>  3:          35+   F          1    13    0    0
#>  4:        15-24   M          1     7    1    0
#>  5:        25-34   M          1    18    0    1
#>  6:          35+   M          1    11    0    0
#>  7:        15-24   F          2    12    1    0
#>  8:        25-34   F          2    24    0    1
#>  9:          35+   F          2    12    0    0
#> 10:        15-24   M          2     6    1    0
#> 11:        25-34   M          2    13    0    1
#> 12:          35+   M          2    13    0    0

dg[, MALE:= as.integer(SEX=='M')]

#>     AGE_AT_MID_C SEX COMM_NUM_B TRIAL AGE1 AGE2 MALE
#>  1:        15-24   F          1    21    1    0    0
#>  2:        25-34   F          1    50    0    1    0
#>  3:          35+   F          1    13    0    0    0
#>  4:        15-24   M          1     7    1    0    1
#>  5:        25-34   M          1    18    0    1    1
#>  6:          35+   M          1    11    0    0    1
#>  7:        15-24   F          2    12    1    0    0
#>  8:        25-34   F          2    24    0    1    0
#>  9:          35+   F          2    12    0    0    0
#> 10:        15-24   M          2     6    1    0    1
#> 11:        25-34   M          2    13    0    1    1
#> 12:          35+   M          2    13    0    0    1

dg[,LOGIT_P_SEQ:=a + comm[COMM_NUM_B] + male*MALE + young*AGE1 + midage*AGE2]

#>     AGE_AT_MID_C SEX COMM_NUM_B TRIAL AGE1 AGE2 MALE LOGIT_P_SEQ
#>  1:        15-24   F          1    21    1    0    0         1.2
#>  2:        25-34   F          1    50    0    1    0         1.1
#>  3:          35+   F          1    13    0    0    0         0.9
#>  4:        15-24   M          1     7    1    0    1         1.6
#>  5:        25-34   M          1    18    0    1    1         1.5
#>  6:          35+   M          1    11    0    0    1         1.3
#>  7:        15-24   F          2    12    1    0    0         0.2
#>  8:        25-34   F          2    24    0    1    0         0.1
#>  9:          35+   F          2    12    0    0    0        -0.1
#> 10:        15-24   M          2     6    1    0    1         0.6
#> 11:        25-34   M          2    13    0    1    1         0.5
#> 12:          35+   M          2    13    0    0    1         0.3

dg[,P_SEQ:=exp(LOGIT_P_SEQ)/(1+exp(LOGIT_P_SEQ))]

#>     AGE_AT_MID_C SEX COMM_NUM_B TRIAL AGE1 AGE2 MALE LOGIT_P_SEQ     P_SEQ
#>  1:        15-24   F          1    21    1    0    0         1.2 0.7685248
#>  2:        25-34   F          1    50    0    1    0         1.1 0.7502601
#>  3:          35+   F          1    13    0    0    0         0.9 0.7109495
#>  4:        15-24   M          1     7    1    0    1         1.6 0.8320184
#>  5:        25-34   M          1    18    0    1    1         1.5 0.8175745
#>  6:          35+   M          1    11    0    0    1         1.3 0.7858350
#>  7:        15-24   F          2    12    1    0    0         0.2 0.5498340
#>  8:        25-34   F          2    24    0    1    0         0.1 0.5249792
#>  9:          35+   F          2    12    0    0    0        -0.1 0.4750208
#> 10:        15-24   M          2     6    1    0    1         0.6 0.6456563
#> 11:        25-34   M          2    13    0    1    1         0.5 0.6224593
#> 12:          35+   M          2    13    0    0    1         0.3 0.5744425

dg[,SUC:=rbinom(nrow(dg),TRIAL,P_SEQ)]

#>     AGE_AT_MID_C SEX COMM_NUM_B TRIAL AGE1 AGE2 MALE LOGIT_P_SEQ     P_SEQ SUC
#>  1:        15-24   F          1    21    1    0    0         1.2 0.7685248  17
#>  2:        25-34   F          1    50    0    1    0         1.1 0.7502601  35
#>  3:          35+   F          1    13    0    0    0         0.9 0.7109495  10
#>  4:        15-24   M          1     7    1    0    1         1.6 0.8320184   5
#>  5:        25-34   M          1    18    0    1    1         1.5 0.8175745  12
#>  6:          35+   M          1    11    0    0    1         1.3 0.7858350  11
#>  7:        15-24   F          2    12    1    0    0         0.2 0.5498340   6
#>  8:        25-34   F          2    24    0    1    0         0.1 0.5249792  10
#>  9:          35+   F          2    12    0    0    0        -0.1 0.4750208   6
#> 10:        15-24   M          2     6    1    0    1         0.6 0.6456563   4
#> 11:        25-34   M          2    13    0    1    1         0.5 0.6224593   5
#> 12:          35+   M          2    13    0    0    1         0.3 0.5744425   8

dg[,CATEGORY:=paste0(COMM_NUM_B,':',SEX,':',AGE_AT_MID_C)]

#>     AGE_AT_MID_C SEX COMM_NUM_B TRIAL AGE1 AGE2 MALE LOGIT_P_SEQ     P_SEQ SUC
#>  1:        15-24   F          1    21    1    0    0         1.2 0.7685248  17
#>  2:        25-34   F          1    50    0    1    0         1.1 0.7502601  35
#>  3:          35+   F          1    13    0    0    0         0.9 0.7109495  10
#>  4:        15-24   M          1     7    1    0    1         1.6 0.8320184   5
#>  5:        25-34   M          1    18    0    1    1         1.5 0.8175745  12
#>  6:          35+   M          1    11    0    0    1         1.3 0.7858350  11
#>  7:        15-24   F          2    12    1    0    0         0.2 0.5498340   6
#>  8:        25-34   F          2    24    0    1    0         0.1 0.5249792  10
#>  9:          35+   F          2    12    0    0    0        -0.1 0.4750208   6
#> 10:        15-24   M          2     6    1    0    1         0.6 0.6456563   4
#> 11:        25-34   M          2    13    0    1    1         0.5 0.6224593   5
#> 12:          35+   M          2    13    0    0    1         0.3 0.5744425   8
#>      CATEGORY
#>  1: 1:F:15-24
#>  2: 1:F:25-34
#>  3:   1:F:35+
#>  4: 1:M:15-24
#>  5: 1:M:25-34
#>  6:   1:M:35+
#>  7: 2:F:15-24
#>  8: 2:F:25-34
#>  9:   2:F:35+
#> 10: 2:M:15-24
#> 11: 2:M:25-34
#> 12:   2:M:35+

dg[,COMM_NUM_B:=as.integer(COMM_NUM_B)]

#>     AGE_AT_MID_C SEX COMM_NUM_B TRIAL AGE1 AGE2 MALE LOGIT_P_SEQ     P_SEQ SUC
#>  1:        15-24   F          1    21    1    0    0         1.2 0.7685248  17
#>  2:        25-34   F          1    50    0    1    0         1.1 0.7502601  35
#>  3:          35+   F          1    13    0    0    0         0.9 0.7109495  10
#>  4:        15-24   M          1     7    1    0    1         1.6 0.8320184   5
#>  5:        25-34   M          1    18    0    1    1         1.5 0.8175745  12
#>  6:          35+   M          1    11    0    0    1         1.3 0.7858350  11
#>  7:        15-24   F          2    12    1    0    0         0.2 0.5498340   6
#>  8:        25-34   F          2    24    0    1    0         0.1 0.5249792  10
#>  9:          35+   F          2    12    0    0    0        -0.1 0.4750208   6
#> 10:        15-24   M          2     6    1    0    1         0.6 0.6456563   4
#> 11:        25-34   M          2    13    0    1    1         0.5 0.6224593   5
#> 12:          35+   M          2    13    0    0    1         0.3 0.5744425   8
#>      CATEGORY
#>  1: 1:F:15-24
#>  2: 1:F:25-34
#>  3:   1:F:35+
#>  4: 1:M:15-24
#>  5: 1:M:25-34
#>  6:   1:M:35+
#>  7: 2:F:15-24
#>  8: 2:F:25-34
#>  9:   2:F:35+
#> 10: 2:M:15-24
#> 11: 2:M:25-34
#> 12:   2:M:35+

#dc
dc<-as.data.table(expand.grid(TR_CATEGORY=dg$CATEGORY,REC_CATEGORY=dg$CATEGORY))
dc[,COUNT_ID:=1:nrow(dc)]

#>      TR_CATEGORY REC_CATEGORY COUNT_ID
#>   1:   1:F:15-24    1:F:15-24        1
#>   2:   1:F:25-34    1:F:15-24        2
#>   3:     1:F:35+    1:F:15-24        3
#>   4:   1:M:15-24    1:F:15-24        4
#>   5:   1:M:25-34    1:F:15-24        5
#>  ---                                  
#> 140:   2:F:25-34      2:M:35+      140
#> 141:     2:F:35+      2:M:35+      141
#> 142:   2:M:15-24      2:M:35+      142
#> 143:   2:M:25-34      2:M:35+      143
#> 144:     2:M:35+      2:M:35+      144

dc[, c("TR_COMM_NUM_B","TR_SEX", "TR_AGE_AT_MID_C") := tstrsplit(TR_CATEGORY, ":", fixed=TRUE)]

#>      TR_CATEGORY REC_CATEGORY COUNT_ID TR_COMM_NUM_B TR_SEX TR_AGE_AT_MID_C
#>   1:   1:F:15-24    1:F:15-24        1             1      F           15-24
#>   2:   1:F:25-34    1:F:15-24        2             1      F           25-34
#>   3:     1:F:35+    1:F:15-24        3             1      F             35+
#>   4:   1:M:15-24    1:F:15-24        4             1      M           15-24
#>   5:   1:M:25-34    1:F:15-24        5             1      M           25-34
#>  ---                                                                       
#> 140:   2:F:25-34      2:M:35+      140             2      F           25-34
#> 141:     2:F:35+      2:M:35+      141             2      F             35+
#> 142:   2:M:15-24      2:M:35+      142             2      M           15-24
#> 143:   2:M:25-34      2:M:35+      143             2      M           25-34
#> 144:     2:M:35+      2:M:35+      144             2      M             35+

dc[, c("REC_COMM_NUM_B","REC_SEX", "REC_AGE_AT_MID_C") := tstrsplit(REC_CATEGORY, ":", fixed=TRUE)]

#>      TR_CATEGORY REC_CATEGORY COUNT_ID TR_COMM_NUM_B TR_SEX TR_AGE_AT_MID_C
#>   1:   1:F:15-24    1:F:15-24        1             1      F           15-24
#>   2:   1:F:25-34    1:F:15-24        2             1      F           25-34
#>   3:     1:F:35+    1:F:15-24        3             1      F             35+
#>   4:   1:M:15-24    1:F:15-24        4             1      M           15-24
#>   5:   1:M:25-34    1:F:15-24        5             1      M           25-34
#>  ---                                                                       
#> 140:   2:F:25-34      2:M:35+      140             2      F           25-34
#> 141:     2:F:35+      2:M:35+      141             2      F             35+
#> 142:   2:M:15-24      2:M:35+      142             2      M           15-24
#> 143:   2:M:25-34      2:M:35+      143             2      M           25-34
#> 144:     2:M:35+      2:M:35+      144             2      M             35+
#>      REC_COMM_NUM_B REC_SEX REC_AGE_AT_MID_C
#>   1:              1       F            15-24
#>   2:              1       F            15-24
#>   3:              1       F            15-24
#>   4:              1       F            15-24
#>   5:              1       F            15-24
#>  ---                                        
#> 140:              2       M              35+
#> 141:              2       M              35+
#> 142:              2       M              35+
#> 143:              2       M              35+
#> 144:              2       M              35+

dc<-dc[TR_SEX!=REC_SEX,]

# transmission proportions
lbd<-matrix(NA_real_,nrow = 1, ncol = nrow(dc))
for (i in 1:nrow(dc)){
  if (dc$TR_COMM_NUM_B[i]==dc$REC_COMM_NUM_B[i]){
    if ( (dc$TR_AGE_AT_MID_C[i]=='15-24' & dc$REC_AGE_AT_MID_C[i]=='25-34') |
         (dc$TR_AGE_AT_MID_C[i]=='15-24' & dc$REC_AGE_AT_MID_C[i]=='35+') |
         (dc$TR_AGE_AT_MID_C[i]=='25-34' & dc$REC_AGE_AT_MID_C[i]=='35+')){
      if (dc$TR_SEX[i]=='M'){
        lbd[i]<-rnorm(1,1,0.2)
      }else{
        lbd[i]<-rnorm(1,2,0.2)
      }
    }else if ( (dc$REC_AGE_AT_MID_C[i]=='15-24' & dc$TR_AGE_AT_MID_C[i]=='25-34') |
               (dc$REC_AGE_AT_MID_C[i]=='15-24' & dc$TR_AGE_AT_MID_C[i]=='35+') |
               (dc$REC_AGE_AT_MID_C[i]=='25-34' & dc$TR_AGE_AT_MID_C[i]=='35+')){
      if (dc$TR_SEX[i]=='M'){
        lbd[i]<-rnorm(1,2,0.2)
      }else{
        lbd[i]<-rnorm(1,1,0.2)
      }
    }else if(dc$TR_AGE_AT_MID_C[i]==dc$REC_AGE_AT_MID_C[i]){
      lbd[i]<-rnorm(1,2,0.2)
    }
  }else{
    lbd[i]<-rnorm(1,1,0.2)
  }
}
TRUE_PI<-rdirichlet(1,lbd)

tmp<-subset(dg,select=c('CATEGORY','P_SEQ'))
setnames(tmp,colnames(tmp),paste0('TR_',colnames(tmp)))
dc<-merge(dc,tmp,by='TR_CATEGORY')
setnames(tmp,colnames(tmp),gsub('TR_','REC_',colnames(tmp)))
dc<-merge(dc,tmp,by='REC_CATEGORY')
dc[,S:=TR_P_SEQ*REC_P_SEQ]

#>     REC_CATEGORY TR_CATEGORY COUNT_ID TR_COMM_NUM_B TR_SEX TR_AGE_AT_MID_C
#>  1:    1:F:15-24   1:M:15-24        4             1      M           15-24
#>  2:    1:F:15-24   1:M:25-34        5             1      M           25-34
#>  3:    1:F:15-24     1:M:35+        6             1      M             35+
#>  4:    1:F:15-24   2:M:15-24       10             2      M           15-24
#>  5:    1:F:15-24   2:M:25-34       11             2      M           25-34
#>  6:    1:F:15-24     2:M:35+       12             2      M             35+
#>  7:    1:F:25-34   1:M:15-24       16             1      M           15-24
#>  8:    1:F:25-34   1:M:25-34       17             1      M           25-34
#>  9:    1:F:25-34     1:M:35+       18             1      M             35+
#> 10:    1:F:25-34   2:M:15-24       22             2      M           15-24
#> 11:    1:F:25-34   2:M:25-34       23             2      M           25-34
#> 12:    1:F:25-34     2:M:35+       24             2      M             35+
#> 13:      1:F:35+   1:M:15-24       28             1      M           15-24
#> 14:      1:F:35+   1:M:25-34       29             1      M           25-34
#> 15:      1:F:35+     1:M:35+       30             1      M             35+
#> 16:      1:F:35+   2:M:15-24       34             2      M           15-24
#> 17:      1:F:35+   2:M:25-34       35             2      M           25-34
#> 18:      1:F:35+     2:M:35+       36             2      M             35+
#> 19:    1:M:15-24   1:F:15-24       37             1      F           15-24
#> 20:    1:M:15-24   1:F:25-34       38             1      F           25-34
#> 21:    1:M:15-24     1:F:35+       39             1      F             35+
#> 22:    1:M:15-24   2:F:15-24       43             2      F           15-24
#> 23:    1:M:15-24   2:F:25-34       44             2      F           25-34
#> 24:    1:M:15-24     2:F:35+       45             2      F             35+
#> 25:    1:M:25-34   1:F:15-24       49             1      F           15-24
#> 26:    1:M:25-34   1:F:25-34       50             1      F           25-34
#> 27:    1:M:25-34     1:F:35+       51             1      F             35+
#> 28:    1:M:25-34   2:F:15-24       55             2      F           15-24
#> 29:    1:M:25-34   2:F:25-34       56             2      F           25-34
#> 30:    1:M:25-34     2:F:35+       57             2      F             35+
#> 31:      1:M:35+   1:F:15-24       61             1      F           15-24
#> 32:      1:M:35+   1:F:25-34       62             1      F           25-34
#> 33:      1:M:35+     1:F:35+       63             1      F             35+
#> 34:      1:M:35+   2:F:15-24       67             2      F           15-24
#> 35:      1:M:35+   2:F:25-34       68             2      F           25-34
#> 36:      1:M:35+     2:F:35+       69             2      F             35+
#> 37:    2:F:15-24   1:M:15-24       76             1      M           15-24
#> 38:    2:F:15-24   1:M:25-34       77             1      M           25-34
#> 39:    2:F:15-24     1:M:35+       78             1      M             35+
#> 40:    2:F:15-24   2:M:15-24       82             2      M           15-24
#> 41:    2:F:15-24   2:M:25-34       83             2      M           25-34
#> 42:    2:F:15-24     2:M:35+       84             2      M             35+
#> 43:    2:F:25-34   1:M:15-24       88             1      M           15-24
#> 44:    2:F:25-34   1:M:25-34       89             1      M           25-34
#> 45:    2:F:25-34     1:M:35+       90             1      M             35+
#> 46:    2:F:25-34   2:M:15-24       94             2      M           15-24
#> 47:    2:F:25-34   2:M:25-34       95             2      M           25-34
#> 48:    2:F:25-34     2:M:35+       96             2      M             35+
#> 49:      2:F:35+   1:M:15-24      100             1      M           15-24
#> 50:      2:F:35+   1:M:25-34      101             1      M           25-34
#> 51:      2:F:35+     1:M:35+      102             1      M             35+
#> 52:      2:F:35+   2:M:15-24      106             2      M           15-24
#> 53:      2:F:35+   2:M:25-34      107             2      M           25-34
#> 54:      2:F:35+     2:M:35+      108             2      M             35+
#> 55:    2:M:15-24   1:F:15-24      109             1      F           15-24
#> 56:    2:M:15-24   1:F:25-34      110             1      F           25-34
#> 57:    2:M:15-24     1:F:35+      111             1      F             35+
#> 58:    2:M:15-24   2:F:15-24      115             2      F           15-24
#> 59:    2:M:15-24   2:F:25-34      116             2      F           25-34
#> 60:    2:M:15-24     2:F:35+      117             2      F             35+
#> 61:    2:M:25-34   1:F:15-24      121             1      F           15-24
#> 62:    2:M:25-34   1:F:25-34      122             1      F           25-34
#> 63:    2:M:25-34     1:F:35+      123             1      F             35+
#> 64:    2:M:25-34   2:F:15-24      127             2      F           15-24
#> 65:    2:M:25-34   2:F:25-34      128             2      F           25-34
#> 66:    2:M:25-34     2:F:35+      129             2      F             35+
#> 67:      2:M:35+   1:F:15-24      133             1      F           15-24
#> 68:      2:M:35+   1:F:25-34      134             1      F           25-34
#> 69:      2:M:35+     1:F:35+      135             1      F             35+
#> 70:      2:M:35+   2:F:15-24      139             2      F           15-24
#> 71:      2:M:35+   2:F:25-34      140             2      F           25-34
#> 72:      2:M:35+     2:F:35+      141             2      F             35+
#>     REC_CATEGORY TR_CATEGORY COUNT_ID TR_COMM_NUM_B TR_SEX TR_AGE_AT_MID_C
#>     REC_COMM_NUM_B REC_SEX REC_AGE_AT_MID_C  TR_P_SEQ REC_P_SEQ         S
#>  1:              1       F            15-24 0.8320184 0.7685248 0.6394267
#>  2:              1       F            15-24 0.8175745 0.7685248 0.6283262
#>  3:              1       F            15-24 0.7858350 0.7685248 0.6039337
#>  4:              1       F            15-24 0.6456563 0.7685248 0.4962029
#>  5:              1       F            15-24 0.6224593 0.7685248 0.4783754
#>  6:              1       F            15-24 0.5744425 0.7685248 0.4414733
#>  7:              1       F            25-34 0.8320184 0.7502601 0.6242302
#>  8:              1       F            25-34 0.8175745 0.7502601 0.6133935
#>  9:              1       F            25-34 0.7858350 0.7502601 0.5895806
#> 10:              1       F            25-34 0.6456563 0.7502601 0.4844102
#> 11:              1       F            25-34 0.6224593 0.7502601 0.4670064
#> 12:              1       F            25-34 0.5744425 0.7502601 0.4309813
#> 13:              1       F              35+ 0.8320184 0.7109495 0.5915231
#> 14:              1       F              35+ 0.8175745 0.7109495 0.5812542
#> 15:              1       F              35+ 0.7858350 0.7109495 0.5586890
#> 16:              1       F              35+ 0.6456563 0.7109495 0.4590290
#> 17:              1       F              35+ 0.6224593 0.7109495 0.4425372
#> 18:              1       F              35+ 0.5744425 0.7109495 0.4083996
#> 19:              1       M            15-24 0.7685248 0.8320184 0.6394267
#> 20:              1       M            15-24 0.7502601 0.8320184 0.6242302
#> 21:              1       M            15-24 0.7109495 0.8320184 0.5915231
#> 22:              1       M            15-24 0.5498340 0.8320184 0.4574720
#> 23:              1       M            15-24 0.5249792 0.8320184 0.4367923
#> 24:              1       M            15-24 0.4750208 0.8320184 0.3952260
#> 25:              1       M            25-34 0.7685248 0.8175745 0.6283262
#> 26:              1       M            25-34 0.7502601 0.8175745 0.6133935
#> 27:              1       M            25-34 0.7109495 0.8175745 0.5812542
#> 28:              1       M            25-34 0.5498340 0.8175745 0.4495302
#> 29:              1       M            25-34 0.5249792 0.8175745 0.4292096
#> 30:              1       M            25-34 0.4750208 0.8175745 0.3883649
#> 31:              1       M              35+ 0.7685248 0.7858350 0.6039337
#> 32:              1       M              35+ 0.7502601 0.7858350 0.5895806
#> 33:              1       M              35+ 0.7109495 0.7858350 0.5586890
#> 34:              1       M              35+ 0.5498340 0.7858350 0.4320788
#> 35:              1       M              35+ 0.5249792 0.7858350 0.4125470
#> 36:              1       M              35+ 0.4750208 0.7858350 0.3732880
#> 37:              2       F            15-24 0.8320184 0.5498340 0.4574720
#> 38:              2       F            15-24 0.8175745 0.5498340 0.4495302
#> 39:              2       F            15-24 0.7858350 0.5498340 0.4320788
#> 40:              2       F            15-24 0.6456563 0.5498340 0.3550038
#> 41:              2       F            15-24 0.6224593 0.5498340 0.3422493
#> 42:              2       F            15-24 0.5744425 0.5498340 0.3158480
#> 43:              2       F            25-34 0.8320184 0.5249792 0.4367923
#> 44:              2       F            25-34 0.8175745 0.5249792 0.4292096
#> 45:              2       F            25-34 0.7858350 0.5249792 0.4125470
#> 46:              2       F            25-34 0.6456563 0.5249792 0.3389561
#> 47:              2       F            25-34 0.6224593 0.5249792 0.3267782
#> 48:              2       F            25-34 0.5744425 0.5249792 0.3015704
#> 49:              2       F              35+ 0.8320184 0.4750208 0.3952260
#> 50:              2       F              35+ 0.8175745 0.4750208 0.3883649
#> 51:              2       F              35+ 0.7858350 0.4750208 0.3732880
#> 52:              2       F              35+ 0.6456563 0.4750208 0.3067002
#> 53:              2       F              35+ 0.6224593 0.4750208 0.2956811
#> 54:              2       F              35+ 0.5744425 0.4750208 0.2728722
#> 55:              2       M            15-24 0.7685248 0.6456563 0.4962029
#> 56:              2       M            15-24 0.7502601 0.6456563 0.4844102
#> 57:              2       M            15-24 0.7109495 0.6456563 0.4590290
#> 58:              2       M            15-24 0.5498340 0.6456563 0.3550038
#> 59:              2       M            15-24 0.5249792 0.6456563 0.3389561
#> 60:              2       M            15-24 0.4750208 0.6456563 0.3067002
#> 61:              2       M            25-34 0.7685248 0.6224593 0.4783754
#> 62:              2       M            25-34 0.7502601 0.6224593 0.4670064
#> 63:              2       M            25-34 0.7109495 0.6224593 0.4425372
#> 64:              2       M            25-34 0.5498340 0.6224593 0.3422493
#> 65:              2       M            25-34 0.5249792 0.6224593 0.3267782
#> 66:              2       M            25-34 0.4750208 0.6224593 0.2956811
#> 67:              2       M              35+ 0.7685248 0.5744425 0.4414733
#> 68:              2       M              35+ 0.7502601 0.5744425 0.4309813
#> 69:              2       M              35+ 0.7109495 0.5744425 0.4083996
#> 70:              2       M              35+ 0.5498340 0.5744425 0.3158480
#> 71:              2       M              35+ 0.5249792 0.5744425 0.3015704
#> 72:              2       M              35+ 0.4750208 0.5744425 0.2728722
#>     REC_COMM_NUM_B REC_SEX REC_AGE_AT_MID_C  TR_P_SEQ REC_P_SEQ         S

N<-rpois(1,20/mean(dc$S))
z<-rmultinom(1,size=N,prob=TRUE_PI)
n<-matrix(NA_integer_,ncol=1,nrow=length(TRUE_PI))
for (i in 1:length(TRUE_PI)){
  n[i]<-rbinom(1,size=z[i],dc$S[i])
}

dc$OBS<-n
dc.obs.zero<-which(dc$OBS==0)
dc<-dc[which(dc$OBS!=0),]
dc$COUNT_ID<-1:nrow(dc)

# remove irrelevant terms
dg$P_SEQ<-NULL
dg$LOGIT_P_SEQ<-NULL
dc$TR_P_SEQ<-NULL
dc$REC_P_SEQ<-NULL
dc$S<-NULL

df.sampling<-list()
df.sampling[[1]]<-dg

samples.from.GLM.prior(df.sampling=df.sampling,dc=dc,iteration=1e6,sample.method='empirical',glm.model='binomial',seed=123)

#> Loading required package: ggplot2

#> Loading required package: StanHeaders

#> Warning: package ‘StanHeaders’ was built under R version 3.5.2

#> rstan (Version 2.18.2, GitRev: 2e1f913d3ca3)

#> For execution on a local, multicore CPU with excess RAM we recommend calling
#> options(mc.cores = parallel::detectCores()).
#> To avoid recompilation of unchanged Stan programs, we recommend calling
#> rstan_options(auto_write = TRUE)

#> 
#> Attaching package: ‘rstan’

#> The following object is masked from ‘package:coda’:
#> 
#>     traceplot

#> Warning: package ‘fitdistrplus’ was built under R version 3.5.2

#> Loading required package: survival

#> Loading required package: npsurv

#> Loading required package: lsei

#> Error in is.data.table(y): object 'dg' not found